organizing files

+++ app/src/main/java/com/hamedaasi/weathernotifyer/api_fetch/weather/CurrentTemperature.kt
@@ -3,5 +3,5 @@ package com.hamedaasi.weathernotifyer.api_fetch.weather
 data class CurrentTemperature(
     val time : String,
     val interval : Int,
-    val temperature_2m : Int,
+    val temperature_2m : Float,
 )
